move test to art commands
